+++ tests/testthat/test-reduce.R
@@ -1,5 +1,32 @@
 context("reduce")
 
+test_that("empty input returns init or error", {
+  expect_error(reduce(list()), "no `.init` supplied")
+  expect_equal(reduce(list(), `+`, .init = 0), 0)
+})
+
+test_that("first/value value used as first value", {
+  x <- c(1, 1)
+
+  expect_equal(reduce(x, `+`), 2)
+  expect_equal(reduce(x, `+`, .init = 1), 3)
+  expect_equal(reduce_right(x, `+`), 2)
+  expect_equal(reduce_right(x, `+`, .init = 1), 3)
+})
+
+test_that("length 1 argument reduced with init", {
+  expect_equal(reduce(1, `+`, .init = 1), 2)
+  expect_equal(reduce_right(1, `+`, .init = 1), 2)
+})
+
+test_that("reduce_right equivalent to reversing input", {
+  x <- list(c(2, 1), c(4, 3), c(6, 5))
+  expect_equal(reduce_right(x, c), c(6, 5, 4, 3, 2, 1))
+  expect_equal(reduce_right(x, c, .init = 7), c(7, 6, 5, 4, 3, 2, 1))
+})
+
+# accumulate --------------------------------------------------------------
+
 test_that("accumulate passes arguments to function", {
   tt <- c("a", "b", "c")
   expect_equal(accumulate(tt, paste, sep = "."), c("a", "a.b", "a.b.c"))